Sanity checks and more commands!
This update brings a lot to the table -- One being that new sanity checks are in place!

If a user is not banned, the new unban command will break and return that they are not banned.

Fixed: Reasons were not being applied to tempmutes and now work again.
Stats would refuse to return anything due to an endpoint mismatch.

Added: Unban, Tempban, playerreport